+++ commands/quarantine.go
@@ -4,10 +4,12 @@ import (
 	"fmt"
 
 	dg "github.com/bwmarrin/discordgo"
+	"github.com/dshoreman/smegbot/config"
 	"github.com/dshoreman/smegbot/util"
 )
 
 func nuke(s *dg.Session, m *dg.MessageCreate) {
+	config.LoadGuild(m.GuildID)
 	sinbin := quarantineRole(s, m.ChannelID, m.GuildID)
 	if sinbin != "" {
 		u := m.Mentions[0].ID
@@ -43,6 +45,9 @@ func restore(s *dg.Session, m *dg.MessageCreate) {
 
 func quarantineRole(s *dg.Session, channelID string, guildID string) string {
 	roles, _ := s.GuildRoles(guildID)
+	if config.Guild.QuarantineRole != "" {
+		return config.Guild.QuarantineRole
+	}
 	for _, r := range roles {
 		if r.Name == "Quarantine" {
 			return r.ID
@@ -80,12 +85,12 @@ func restoreRoles(s *dg.Session, g string, c string, u string) {
 }
 
 func sendSuccessOrFail(s *dg.Session, channelID string, err error, mode string, target string) {
-	op, result := "adding", "now in Quarantine."
+	op, result := "add", "now in Quarantine."
 	if mode == "remove" {
-		op, result = "removing", "back out of Quarantine!"
+		op, result = "remove", "back out of Quarantine!"
 	}
 	if err != nil {
-		s.ChannelMessageSend(channelID, "Oops! Couldn't "+op+" the **@Quarantine** role.")
+		s.ChannelMessageSend(channelID, "Oops! Couldn't "+op+" the configured quarantine role. Check permissions!")
 		fmt.Println("\nError:\n", err)
 		return
 	}