+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [1.3.0] - 2020-06-18
+### Added
+* Quarantine role is now able to be configured per-guild
+
+### Fixed
+* Phrasing of quarantine errors now uses correct tense
+
+
 ## [1.2.1] - 2020-06-18
 ### Changed
 * Message logs in console now include guild and channel names
@@ -48,7 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Global permission checks - messages from members without Administrator are ignored
 
 
-[Unreleased]: https://github.com/dshoreman/smegbot/compare/v1.2.1...develop
+[Unreleased]: https://github.com/dshoreman/smegbot/compare/v1.3.0...develop
+[1.3.0]: https://github.com/dshoreman/smegbot/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/dshoreman/smegbot/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/dshoreman/smegbot/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/dshoreman/smegbot/compare/v1.0.1...v1.1.0